+++ src/main/java/dev/lambdaurora/auroraslanterns/resource/AurorasLanternsStaticDatagen.java
@@ -17,13 +17,19 @@ import net.fabricmc.fabric.api.datagen.v1.FabricDataGenerator;
 import net.fabricmc.fabric.api.datagen.v1.FabricDataOutput;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricAdvancementProvider;
 import net.fabricmc.fabric.api.datagen.v1.provider.FabricBlockLootTableProvider;
+import net.fabricmc.fabric.api.datagen.v1.provider.FabricRecipeProvider;
 import net.minecraft.advancements.Advancement;
 import net.minecraft.advancements.AdvancementHolder;
 import net.minecraft.advancements.AdvancementType;
 import net.minecraft.core.HolderLookup;
+import net.minecraft.data.recipes.RecipeCategory;
+import net.minecraft.data.recipes.RecipeOutput;
+import net.minecraft.data.recipes.RecipeProvider;
 import net.minecraft.network.chat.Text;
 import net.minecraft.resources.Identifier;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.level.block.Blocks;
+import org.jetbrains.annotations.NotNull;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
@@ -34,6 +40,7 @@ public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoi
 		var pack = fabricDataGenerator.createPack();
 		pack.addProvider(LootDataProvider::new);
 		pack.addProvider(AdvancementProvider::new);
+		pack.addProvider(AurorasRecipeProvider.Runner::new);
 	}
 
 	private static class LootDataProvider extends FabricBlockLootTableProvider {
@@ -74,6 +81,50 @@ public final class AurorasLanternsStaticDatagen implements DataGeneratorEntrypoi
 					.addCriterion("bonk", WallLanternBonkTrigger.TriggerInstance.bonk())
 					.build(AurorasLanterns.id("adventure/wall_lantern_bonk"))
 			);
+		}
+	}
+
+	private static class AurorasRecipeProvider extends RecipeProvider {
+		public AurorasRecipeProvider(HolderLookup.Provider provider, RecipeOutput recipeOutput) {
+			super(provider, recipeOutput);
+		}
+
+		@Override
+		public void buildRecipes() {
+			this.shaped(RecipeCategory.DECORATIONS, AurorasLanternsRegistry.AMETHYST_LANTERN_ITEM)
+					.define('S', Items.AMETHYST_SHARD)
+					.define('L', Items.LANTERN)
+					.pattern("SLS")
+					.unlockedBy("has_amethyst_shard", has(Items.AMETHYST_SHARD))
+					.unlockedBy("has_lantern", has(Items.LANTERN))
+					.unlockedBy("has_self", has(AurorasLanternsRegistry.AMETHYST_LANTERN_ITEM))
+					.save(this.output);
+			this.shaped(RecipeCategory.REDSTONE, AurorasLanternsRegistry.REDSTONE_LANTERN_ITEM)
+					.define('I', Items.IRON_NUGGET)
+					.define('T', Items.REDSTONE_TORCH)
+					.pattern("III")
+					.pattern("ITI")
+					.pattern("III")
+					.unlockedBy("has_iron_nugget", has(Items.IRON_NUGGET))
+					.unlockedBy("has_redstone_torch", has(Items.REDSTONE_TORCH))
+					.unlockedBy("has_self", has(AurorasLanternsRegistry.REDSTONE_LANTERN_ITEM))
+					.save(this.output);
+		}
+
+		private static class Runner extends FabricRecipeProvider {
+			public Runner(FabricDataOutput output, CompletableFuture<HolderLookup.Provider> registriesFuture) {
+				super(output, registriesFuture);
+			}
+
+			@Override
+			protected @NotNull RecipeProvider createRecipeProvider(HolderLookup.Provider provider, RecipeOutput recipeOutput) {
+				return new AurorasRecipeProvider(provider, recipeOutput);
+			}
+
+			@Override
+			public @NotNull String getName() {
+				return "Aurora's Lanterns Recipes";
+			}
 		}
 	}
 }