+++ src/main/java/dev/lambdaurora/auroraslanterns/test/RedstoneLanternTest.java
@@ -20,8 +20,8 @@ import net.minecraft.world.level.block.Blocks;
 import net.minecraft.world.level.block.RepeaterBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.block.state.properties.RedstoneSide;
-import org.jetbrains.annotations.NotNull;
+
+import static dev.lambdaurora.auroraslanterns.test.TestHelper.assertBlockState;
 
 /**
  * Tests the Redstone Lantern behavior.
@@ -47,15 +47,15 @@ public class RedstoneLanternTest {
 			assertBlockState(context, UNLIT_REDSTONE_LAMP, BlockPos.ofFloored(5, 2, 2));
 
 			// Repeaters
-			assertBlockState(context, litRepeater(Direction.NORTH), BlockPos.ofFloored(1, 1, 3));
-			assertBlockState(context, litRepeater(Direction.EAST), BlockPos.ofFloored(0, 1, 2));
-			assertBlockState(context, litRepeater(Direction.SOUTH), BlockPos.ofFloored(1, 1, 1));
-			assertBlockState(context, litRepeater(Direction.WEST), BlockPos.ofFloored(2, 1, 2));
+			assertBlockState(context, litRepeater(context, Direction.NORTH), BlockPos.ofFloored(1, 1, 3));
+			assertBlockState(context, litRepeater(context, Direction.EAST), BlockPos.ofFloored(0, 1, 2));
+			assertBlockState(context, litRepeater(context, Direction.SOUTH), BlockPos.ofFloored(1, 1, 1));
+			assertBlockState(context, litRepeater(context, Direction.WEST), BlockPos.ofFloored(2, 1, 2));
 
-			assertBlockState(context, litRepeater(Direction.NORTH), BlockPos.ofFloored(5, 1, 3));
-			assertBlockState(context, litRepeater(Direction.EAST), BlockPos.ofFloored(4, 1, 2));
-			assertBlockState(context, litRepeater(Direction.SOUTH), BlockPos.ofFloored(5, 1, 1));
-			assertBlockState(context, litRepeater(Direction.WEST), BlockPos.ofFloored(6, 1, 2));
+			assertBlockState(context, litRepeater(context, Direction.NORTH), BlockPos.ofFloored(5, 1, 3));
+			assertBlockState(context, litRepeater(context, Direction.EAST), BlockPos.ofFloored(4, 1, 2));
+			assertBlockState(context, litRepeater(context, Direction.SOUTH), BlockPos.ofFloored(5, 1, 1));
+			assertBlockState(context, litRepeater(context, Direction.WEST), BlockPos.ofFloored(6, 1, 2));
 
 			context.succeed();
 		});
@@ -64,28 +64,13 @@ public class RedstoneLanternTest {
 	@GameTest(structure = PREFIX + "wall_connectivity")
 	public void testWallConnectivity(GameTestHelper context) {
 		context.runAtTickTime(1, () -> {
-			BlockState xRedstone = Blocks.REDSTONE_WIRE.defaultState()
-					.with(BlockStateProperties.POWER, 15)
-					.with(BlockStateProperties.EAST_REDSTONE, RedstoneSide.SIDE)
-					.with(BlockStateProperties.WEST_REDSTONE, RedstoneSide.SIDE)
-					.with(BlockStateProperties.NORTH_REDSTONE, RedstoneSide.NONE)
-					.with(BlockStateProperties.SOUTH_REDSTONE, RedstoneSide.NONE);
+			assertRedstoneSignal(context, BlockPos.ofFloored(2, 1, 1), Direction.WEST);
+			assertRedstoneSignal(context, BlockPos.ofFloored(0, 1, 1), Direction.EAST);
+			assertRedstoneSignal(context, BlockPos.ofFloored(1, 1, 0), Direction.SOUTH);
 
 			assertBlockState(context, LIT_REDSTONE_LAMP, BlockPos.ofFloored(1, 0, 1));
 			assertBlockState(context, LIT_REDSTONE_LAMP, BlockPos.ofFloored(1, 2, 1));
 			assertBlockState(context, UNLIT_REDSTONE_LAMP, BlockPos.ofFloored(1, 1, 2));
-
-			assertBlockState(context, xRedstone, BlockPos.ofFloored(0, 1, 1));
-			assertBlockState(context, xRedstone, BlockPos.ofFloored(2, 1, 1));
-			assertBlockState(context,
-					Blocks.REDSTONE_WIRE.defaultState()
-							.with(BlockStateProperties.POWER, 15)
-							.with(BlockStateProperties.NORTH_REDSTONE, RedstoneSide.SIDE)
-							.with(BlockStateProperties.SOUTH_REDSTONE, RedstoneSide.SIDE)
-							.with(BlockStateProperties.EAST_REDSTONE, RedstoneSide.NONE)
-							.with(BlockStateProperties.WEST_REDSTONE, RedstoneSide.NONE),
-					BlockPos.ofFloored(1, 1, 0)
-			);
 
 			context.succeed();
 		});
@@ -111,29 +96,21 @@ public class RedstoneLanternTest {
 		});
 	}
 
-	private static BlockState litRepeater(Direction facing) {
+	private static BlockState litRepeater(GameTestHelper context, Direction facing) {
 		return Blocks.REPEATER.defaultState()
 				.with(BlockStateProperties.POWERED, true)
 				.with(BlockStateProperties.LOCKED, false)
-				.with(RepeaterBlock.FACING, facing)
+				.with(RepeaterBlock.FACING, context.getTestRotation().rotate(facing))
 				.with(BlockStateProperties.DELAY, 1);
 	}
 
-	/**
-	 * Expects the given block state at the given block position.
-	 *
-	 * @param state the expected block state
-	 * @param pos   the position to check for
-	 */
-	public static void assertBlockState(
-			GameTestHelper context, @NotNull BlockState state, @NotNull BlockPos pos
-	) {
-		context.assertBlockState(
+	private static void assertRedstoneSignal(GameTestHelper context, BlockPos pos, Direction direction) {
+		var rotated = context.getTestRotation().rotate(direction);
+		context.assertRedstoneSignal(
 				pos,
-				s -> s.equals(state),
-				s -> Text.literal(
-						"Expected block state %s at position %s, found %s.".formatted(state, pos.toShortString(), s)
-				)
+				rotated,
+				value -> value == 15,
+				() -> Text.literal("Missing redstone signal at %s towards %s!".formatted(pos, rotated))
 		);
 	}
 }