+++ CHANGELOG.md
@@ -19,11 +19,30 @@
 - Added better integration of Adorn's Candlelit Lanterns.
 - Fixed some cases where block state IDs would not be able to sync back up.
 
+### 1.0.2
+
+- Improved support of oxidizing and waxable lanterns.
+- Improved wall lanterns on walls that do not have the post part.
+
 ## 1.1.0
 
 - Ported to Minecraft 1.21.1.
 - Improved stability of Aurora's Decorations backwards compatibility.
 
+### 1.1.1
+
+- Improved support of oxidizing and waxable lanterns.
+- Improved wall lanterns on walls that do not have the post part.
+- Fixed missing recipes for the Amethyst and Redstone Lanterns.
+- Fixed missing `#auroraslanterns:wall_lanterns` block tag.
+
 ## 1.2.0
 
 - Ported to Minecraft 1.21.5.
+
+### 1.2.1
+
+- Improved support of oxidizing and waxable lanterns.
+- Improved wall lanterns on walls that do not have the post part.
+- Fixed missing recipes for the Amethyst and Redstone Lanterns.
+- Fixed missing `#auroraslanterns:wall_lanterns` block tag.